size_t to int cast.

+++ src/console_gui.cpp
@@ -68,7 +68,7 @@ console_gui::whitespace(const size_t &size) {
 std::string
 console_gui::center(const std::string text, size_t width)
 {
-	int text_position = (width - text.size()) / 2;
+	int text_position = (int)(width - text.size()) / 2;
 	std::string whitespace = console_gui::whitespace(text_position);
 	std::string centered_text = whitespace + text + whitespace;
 	if (centered_text.size() < width) {
